Map and player spawner
There is a new map ui that alows the player to choose the phase they want to fight and also the character they want to use.

A lot of new scripts were created to hold informations, such as BossDescriptionSO - to hold information of bosses phases - and CharactersSO - to hold information of characters, like name, initial kit and a dictionary of sprites for future use.

The map is in the Menu scene but will later be in the Tavern scene.

Temporary icons for the map were added to the project.

The character the player chooses to use will now be instantiated in the fight they choose in the map.

WhiteBoard and PlayerWhiteBoard were created as singletons to hold dinamic information that must be accessed by different scenes.